requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.4.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
autopep8==1.5.4
click==7.1.2
faker==4.17.1
flask-cors==3.0.8
flask-jwt-extended==3.24.1
flask-login==0.5.0
flask-migrate==2.5.3
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
jinja2==2.11.2
mako==1.1.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==1.1.1
pep8==1.7.1
pycodestyle==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==1.7.1
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
sqlalchemy==1.3.19
text-unidecode==1.3
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
werkzeug==1.0.1
wtforms==2.3.3
